Ai Integrations/requirements.txt
google-genai>=1.50.0
google-cloud-aiplatform>=1.38.0
google-cloud-texttospeech>=2.16.0
httpx>=0.27.0
numpy>=1.26.0
python-dotenv>=1.0.0
runwayml>=4.8.0
